[VFS] OverlayFileSystem should consistently use the last-added FS first
Rather than checking whether a file exists before running an operation
and using the result of the first FS where it exists, always run
operations on the most recently added FS's first. If there's any error
other than missing file, return it immediately.

Also update CWD handling. The previous implementation attempted to
`setCWD` on all contained `FS` and then would use the `CWD` of the first
FS for `getCWD`, assuming they are all synced. This had a bunch of odd
behaviors:
  1. `setCWD` would fail but all FS up to the failure would have their
     CWD set. `getCWD` would then return the just set CWD, even though
     it "failed"
  2. Their CWD could get out of sync, so the CWD used by each FS
     wouldn't necessarily match the just set CWD. This could result in
     odd behaviour where a relative file is found from a previously-set
     CWD.

Keep a CWD in `OverlayFileSystem` instead and pass the already-absoluted
path down into the contained FS. Use the new `ExposesExternalVFSPath` to
determine whether to override the returned path back to the original or
not.

Differential Revision: https://reviews.llvm.org/D121423